# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Arthur Herron Butchery API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/{orderNumber}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get an order by order number
      parameters:
        - name: orderNumber
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Order detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /payments/pesepay/initiate:
    post:
      operationId: initiatePesepayPayment
      tags: [payments]
      summary: Initiate a Pesepay payment for an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PesepayInitiateInput"
      responses:
        "200":
          description: Payment initiated, redirect the customer to redirectUrl
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PesepayInitiateResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /payments/pesepay/status/{orderNumber}:
    get:
      operationId: getPesepayStatus
      tags: [payments]
      summary: Poll Pesepay for the latest status of an order's payment
      parameters:
        - name: orderNumber
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Current payment status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PesepayStatusResponse"
        "404":
          description: Order not found or payment not yet initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /payments/pesepay/result:
    post:
      operationId: pesepayResultWebhook
      tags: [payments]
      summary: Webhook Pesepay calls with the final encrypted transaction result
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PesepayWebhookPayload"
      responses:
        "200":
          description: Acknowledged
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    OrderItemInput:
      type: object
      required: [productId, productName, quantity, unitPrice, cutOption, note]
      properties:
        productId: { type: integer }
        productName: { type: string }
        quantity: { type: number }
        unitPrice: { type: number }
        cutOption: { type: ["string", "null"] }
        note: { type: ["string", "null"] }
    OrderItem:
      type: object
      required: [id, productId, productName, quantity, unitPrice, cutOption, note, lineTotal]
      properties:
        id: { type: integer }
        productId: { type: integer }
        productName: { type: string }
        quantity: { type: number }
        unitPrice: { type: number }
        cutOption: { type: ["string", "null"] }
        note: { type: ["string", "null"] }
        lineTotal: { type: number }
    OrderInput:
      type: object
      required:
        - fullName
        - phone
        - email
        - fulfillmentType
        - address
        - areaZoneId
        - preferredDate
        - preferredTimeWindow
        - notes
        - items
      properties:
        fullName: { type: string, minLength: 1 }
        phone: { type: string, minLength: 1 }
        email: { type: ["string", "null"] }
        fulfillmentType:
          type: string
          enum: [delivery, pickup]
        address: { type: ["string", "null"] }
        areaZoneId: { type: ["integer", "null"] }
        preferredDate: { type: string }
        preferredTimeWindow: { type: string }
        notes: { type: ["string", "null"] }
        items:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/OrderItemInput"
    Order:
      type: object
      required:
        - id
        - orderNumber
        - fullName
        - phone
        - email
        - fulfillmentType
        - address
        - areaZoneId
        - areaZoneName
        - deliveryFee
        - preferredDate
        - preferredTimeWindow
        - notes
        - status
        - subtotal
        - total
        - createdAt
        - items
      properties:
        id: { type: integer }
        orderNumber: { type: string }
        userId: { type: ["string", "null"] }
        fullName: { type: string }
        phone: { type: string }
        email: { type: ["string", "null"] }
        fulfillmentType:
          type: string
          enum: [delivery, pickup]
        address: { type: ["string", "null"] }
        areaZoneId: { type: ["integer", "null"] }
        areaZoneName: { type: ["string", "null"] }
        deliveryFee: { type: number }
        preferredDate: { type: string }
        preferredTimeWindow: { type: string }
        notes: { type: ["string", "null"] }
        status:
          type: string
          enum: [pending_payment, confirmed, payment_failed]
        subtotal: { type: number }
        total: { type: number }
        paymentReferenceNumber: { type: ["string", "null"] }
        createdAt: { type: string }
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
...
    PesepayInitiateInput:
      type: object
      required: [orderNumber]
      properties:
        orderNumber: { type: string, minLength: 1 }
    PesepayInitiateResponse:
      type: object
      required: [redirectUrl, referenceNumber]
      properties:
        redirectUrl: { type: string }
        referenceNumber: { type: string }
    PesepayStatusResponse:
      type: object
      required: [status, orderStatus]
      properties:
        status: { type: string, description: "Raw Pesepay transaction status" }
        orderStatus:
          type: string
          enum: [pending_payment, confirmed, payment_failed]
    PesepayWebhookPayload:
      type: object
      required: [payload]
      properties:
        payload:
          type: string
          description: Base64-encoded AES-256-CBC encrypted transaction payload
...