# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Arthur Herron Butchery API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderFulfillmentType } from './orderFulfillmentType';
import type { OrderItem } from './orderItem';
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: number;
  orderNumber: string;
  /** @nullable */
  userId?: string | null;
  fullName: string;
  phone: string;
  /** @nullable */
  email: string | null;
  fulfillmentType: OrderFulfillmentType;
  /** @nullable */
  address: string | null;
  /** @nullable */
  areaZoneId: number | null;
  /** @nullable */
  areaZoneName: string | null;
  deliveryFee: number;
  preferredDate: string;
  preferredTimeWindow: string;
  /** @nullable */
  notes: string | null;
  status: OrderStatus;
  subtotal: number;
  total: number;
  /** @nullable */
  paymentReferenceNumber?: string | null;
  createdAt: string;
  items: OrderItem[];
}